requirements.txt
 openai=~1.1.1
 python-dotenv==1.0.0